+++ common/sonicv2connector.cpp
@@ -69,7 +69,7 @@ std::vector<std::string> SonicV2Connector::keys(const std::string& db_name, cons
     return m_dbintf.keys(db_name, pattern, blocking);
 }
 
-std::pair<int64_t, std::vector<std::string>> SonicV2Connector::scan(const std::string& db_name, int64_t cursor, const char *match, uint32_t count)
+std::pair<int, std::vector<std::string>> SonicV2Connector::scan(const std::string& db_name, int cursor, const char *match, uint32_t count)
 {
     return m_dbintf.scan(db_name, cursor, match, count);
 }
